Fix for rendering crash on Windows 10
When creating a new console buffer, use FILE_SHARE_READ and
FILE_SHARE_WRITE for "share mode".

The symptom this addresses in termshark is described here:
https://github.com/gcla/termshark/issues/73.

Under the hood, a call to GetConsoleScreenBufferInfo() succeeds, but
does not fill in the provided consoleInfo structure - and it doesn't
fail on the first call, only after a dozen or so successful calls to the
API. This is only on Windows 10 - Windows 7 seems unaffacted. This
"successful" call tricks my tcell changes into thinking the console has
resized to 0x0.

I found this difference in an MSDN sample here:
https://docs.microsoft.com/en-us/windows/console/reading-and-writing-blocks-of-characters-and-attributes?redirectedfrom=MSDN